+++ apps/client/package.json
@@ -32,8 +32,8 @@
     "@react-aria/label": "^3.4.3",
     "@react-aria/ssr": "^3.4.0",
     "@react-stately/data": "^3.8.0",
-    "@sentry/nextjs": "^7.24.2",
-    "@sentry/tracing": "^7.24.2",
+    "@sentry/nextjs": "^7.21.1",
+    "@sentry/tracing": "^7.21.1",
     "@snailycad/config": "1.36.0",
     "@snailycad/permissions": "1.36.0",
     "@snailycad/schemas": "1.36.0",
