+++ apps/client/src/components/account/AppearanceTab.tsx
@@ -10,7 +10,7 @@ import { useTranslations } from "use-intl";
 import { StatusViewMode, TableActionsAlignment } from "@snailycad/types";
 import { Select } from "components/form/Select";
 import { Button, Loader, SelectField } from "@snailycad/ui";
-import nextConfig from "../../../next.config";
+import nextConfig from "../../../next.config.mjs";
 import type { Sounds } from "lib/server/getAvailableSounds.server";
 import { soundCamelCaseToKebabCase } from "lib/utils";
 import { CaretDownFill } from "react-bootstrap-icons";
