+++ apps/client/src/lib/i18n/getNextI18nConfig.ts
@@ -1,4 +1,4 @@
 export async function getNextI18nConfig() {
-  const nextConfig = (await import("../../../next.config")).default;
-  return nextConfig.i18n;
+  const nextConfig = (await import("../../../next.config.mjs")).i18n;
+  return nextConfig;
 }
